Crossfilter rewrite, Aggregates implementation
the crossfilter2 deals better with array-valued dimensions, so rewrite to make use of that.
crossfiltering multiple dimensions in one filter (ie. [x,y,z]) is working now
This probably slows the crossfilter down (a lot?), but the postgres server is
supposed to deal with those cases.

Filters now aggregate over an arbitrary number of facets
TODO:
 * postprocessing for normalization
 * TimeOrDuration facets in postgres
 * User interface rewrite for TimeOrDuration