BUGFIX: broken lineplot
Weird behaviour for linechart selections and plots due to how we updated plots in views/util.js
We removed datasets from the front, everything shifts one place to the 'left', including the dummy dataset at the back for selections.
This caused animations and autoscaling to go crazy.

Solution: remove from the back