BUGFIX: empty plots on analyze page
When a widget does not have any facets defined
(fi. when you removed the datasets the facets came from),
and you navigate to the analyze page, the widgets started with
editMode = false, resulting in an almost invisible empty white square.

This patch prevents empty white squares by only turning editMode off
when there is something on the chart.